Split post 1416 oversized slices
