+++ app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
@@ -456,54 +456,14 @@ class PromptComposerViewModelTest {
         assertNull(vm.uiState.value.error)
     }
 
-    @Test
-    fun attachFilesStagesChipsAndKeepsDraftClean() = runTest {
-        // Issue #544: staging adds structured chips (file name only) to the
-        // attachments list — the draft text is NOT mutated while composing.
-        val vm = newVm()
-        vm.onDraftChange("Review these")
-
-        vm.attachFiles(count = 2) {
-            Result.success(
-                listOf(
-                    "~/.pocketshell/attachments/host-1/20260601-120000-01-report.txt",
-                    "~/.pocketshell/attachments/host-1/20260601-120000-02-data.csv",
-                ),
-            )
-        }
-        advanceUntilIdle()
-
-        // The draft stays exactly what the user typed — no "Attached files:"
-        // bullets folded in.
-        assertEquals("Review these", vm.uiState.value.draft)
-        assertEquals(
-            listOf(
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = "~/.pocketshell/attachments/host-1/20260601-120000-01-report.txt",
-                    displayName = "20260601-120000-01-report.txt",
-                ),
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = "~/.pocketshell/attachments/host-1/20260601-120000-02-data.csv",
-                    displayName = "20260601-120000-02-data.csv",
-                ),
-            ),
-            vm.uiState.value.attachments,
-        )
-        assertNull(vm.uiState.value.error)
-        assertEquals(
-            PromptComposerViewModel.AttachmentUploadState.Idle,
-            vm.uiState.value.attachmentUpload,
-        )
-    }
-
     // -- Issue #1350: Send clears the draft for EVERY prompt shape --------
     //
     // Class-coverage guard for the maintainer's dogfood annoyance ("the next
     // dictation/typing appends to stale text"). `clearComposerForHandoff`
     // (issue #971 single-representation hand-off) empties the editable draft the
     // instant a Send is dispatched, and it is length/shape-INDEPENDENT. These
-    // pin that invariant across the whole class — short, long/wrapping,
-    // multi-line, and attachment-bearing — so a future change can't
+    // pin that invariant across the common text shapes — short, long/wrapping,
+    // and multi-line — so a future change can't
     // reintroduce a stale draft for one shape (e.g. a wrapped long prompt).
 
     @Test
@@ -529,37 +489,6 @@ class PromptComposerViewModelTest {
         )
     }
 
-    @Test
-    fun sendClearsDraftForAttachmentBearingPrompt() = runTest {
-        val vm = newVm()
-        val sends = mutableListOf<PromptComposerViewModel.SendRequest>()
-        val job = launch { vm.sendRequests.collect { sends += it } }
-        vm.onDraftChange("review these carefully and report anything suspicious you find")
-        vm.attachFiles(count = 2) {
-            Result.success(
-                listOf(
-                    "~/.pocketshell/attachments/host-1/20260601-120000-01-report.txt",
-                    "~/.pocketshell/attachments/host-1/20260601-120000-02-data.csv",
-                ),
-            )
-        }
-        advanceUntilIdle()
-        assertEquals(2, vm.uiState.value.attachments.size)
-
-        vm.requestSend(withEnter = true)
-        advanceUntilIdle()
-
-        // Both the draft text AND the staged attachment tiles clear on hand-off.
-        assertEquals("", vm.uiState.value.draft)
-        assertTrue(vm.uiState.value.attachments.isEmpty())
-        assertTrue(vm.uiState.value.sendInFlight)
-        // The composed prompt that actually went out still carried the clean text
-        // (the paths are folded into the outgoing text at send time, #544).
-        assertEquals(1, sends.size)
-        assertTrue(sends.single().text.contains("review these"))
-        job.cancelAndJoin()
-    }
-
     private suspend fun TestScope.assertSendClearsDraft(prompt: String) {
         val vm = newVm()
         val sends = mutableListOf<PromptComposerViewModel.SendRequest>()
@@ -579,278 +508,6 @@ class PromptComposerViewModelTest {
         assertEquals(1, sends.size)
         assertEquals(prompt, sends.single().text)
         job.cancelAndJoin()
-    }
-
-    @Test
-    fun attachFilesDeDuplicatesRepeatedRemotePaths() = runTest {
-        // Issue #544: re-attaching the same remote path must not double a chip.
-        val vm = newVm()
-        val path = "~/.pocketshell/attachments/host-1/20260601-120000-01-report.txt"
-
-        vm.attachFiles(count = 1) { Result.success(listOf(path)) }
-        advanceUntilIdle()
-        vm.attachFiles(count = 1) { Result.success(listOf(path)) }
-        advanceUntilIdle()
-
-        assertEquals(1, vm.uiState.value.attachments.size)
-        assertEquals(path, vm.uiState.value.attachments.single().remotePath)
-    }
-
-    @Test
-    fun removeAttachmentDropsOnlyThatChipAndKeepsDraft() = runTest {
-        // Issue #544: the chip's × removes just that attachment; the draft
-        // text the user typed is untouched, and removing all chips leaves a
-        // clean prompt.
-        val vm = newVm()
-        vm.onDraftChange("Look at these")
-        val report = "~/.pocketshell/attachments/host-1/20260601-120000-01-report.txt"
-        val data = "~/.pocketshell/attachments/host-1/20260601-120000-02-data.csv"
-
-        vm.attachFiles(count = 2) { Result.success(listOf(report, data)) }
-        advanceUntilIdle()
-
-        vm.removeAttachment(report)
-
-        assertEquals(
-            listOf(
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = data,
-                    displayName = "20260601-120000-02-data.csv",
-                ),
-            ),
-            vm.uiState.value.attachments,
-        )
-        assertEquals("Look at these", vm.uiState.value.draft)
-
-        // Removing the last chip leaves the prompt exactly as typed.
-        vm.removeAttachment(data)
-        assertTrue(vm.uiState.value.attachments.isEmpty())
-        assertEquals("Look at these", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun seedAttachmentPreloadsChipAndKeepsDraftClean() = runTest {
-        // Issue #560: the share-into-session flow stages the shared file via
-        // the #544 mechanic, then seeds the resulting remote path here so the
-        // composer opens with the chip already present and the draft clean.
-        val vm = newVm()
-        vm.onDraftChange("")
-        val path = "~/.pocketshell/attachments/host-7-scratch/20260606-120000-01-shot.png"
-
-        vm.seedAttachment(path)
-
-        assertEquals(
-            listOf(
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = path,
-                    displayName = "20260606-120000-01-shot.png",
-                ),
-            ),
-            vm.uiState.value.attachments,
-        )
-        assertEquals("", vm.uiState.value.draft)
-    }
-
-    @Test
-    fun stagedAttachmentsAndDraftPersistAcrossComposerCloseAndSessionSwitch() = runTest {
-        // Issue #673: attachments are no longer shown in the session/terminal
-        // bottom area — they live only inside the Prompt Composer sheet. The
-        // maintainer's desired workflow ("insert prompt → close composer →
-        // switch session → send elsewhere") relies on the staged-attachment
-        // STATE surviving the composer being dismissed and the session being
-        // switched. The composer ViewModel is retained across both, so there is
-        // no clear-on-close path: closing the sheet and switching sessions
-        // must leave the draft + attachments intact so re-opening shows them.
-        val vm = newVm()
-        val first = "~/.pocketshell/attachments/host-1/20260610-120000-01-shot.png"
-        val second = "~/.pocketshell/attachments/host-1/20260610-120000-02-data.csv"
-        vm.onDraftChange("deploy the staging build")
-        vm.seedAttachment(first)
-        vm.seedAttachment(second)
-
-        // Snapshot the staged state as it is when the user closes the sheet.
-        val afterCompose = vm.uiState.value
-
-        // Dismissing the composer sheet and switching sessions does NOT touch
-        // the ViewModel state — no clear-on-close, no clear-on-switch hook. The
-        // only thing that clears attachments is an explicit send. Re-reading the
-        // state (as a re-opened sheet would) shows the same draft + chips.
-        val afterReopen = vm.uiState.value
-        assertEquals("deploy the staging build", afterReopen.draft)
-        assertEquals(
-            listOf(
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = first,
-                    displayName = "20260610-120000-01-shot.png",
-                ),
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = second,
-                    displayName = "20260610-120000-02-data.csv",
-                ),
-            ),
-            afterReopen.attachments,
-        )
-        assertEquals(afterCompose.draft, afterReopen.draft)
-        assertEquals(afterCompose.attachments, afterReopen.attachments)
-    }
-
-    @Test
-    fun seedAttachmentDeDuplicatesAndIgnoresBlank() = runTest {
-        val vm = newVm()
-        val path = "~/.pocketshell/attachments/host-7-scratch/shot.png"
-
-        vm.seedAttachment(path)
-        vm.seedAttachment(path)
-        vm.seedAttachment("   ")
-
-        assertEquals(1, vm.uiState.value.attachments.size)
-        assertEquals(path, vm.uiState.value.attachments.single().remotePath)
-    }
-
-    @Test
-    fun removeAttachmentForUnknownPathIsNoOp() = runTest {
-        val vm = newVm()
-        val path = "~/.pocketshell/attachments/host-1/20260601-120000-01-report.txt"
-        vm.attachFiles(count = 1) { Result.success(listOf(path)) }
-        advanceUntilIdle()
-
-        vm.removeAttachment("~/.pocketshell/attachments/host-1/does-not-exist.txt")
-
-        assertEquals(1, vm.uiState.value.attachments.size)
-        assertEquals(path, vm.uiState.value.attachments.single().remotePath)
-    }
-
-    @Test
-    fun attachFilesFailureKeepsDraftAndShowsError() = runTest {
-        val vm = newVm()
-        vm.onDraftChange("Do not lose this draft")
-
-        vm.attachFiles(count = 1) {
-            Result.failure(IllegalStateException("Permission denied"))
-        }
-        advanceUntilIdle()
-
-        assertEquals("Do not lose this draft", vm.uiState.value.draft)
-        assertEquals(
-            PromptComposerViewModel.AttachmentUploadState.Idle,
-            vm.uiState.value.attachmentUpload,
-        )
-        val error = vm.uiState.value.error
-        assertNotNull(error)
-        assertTrue(error!!.contains("Attachment upload failed"))
-        assertTrue(error.contains("Permission denied"))
-        assertTrue(error.contains("draft was kept"))
-    }
-
-    @Test
-    fun attachFilesPartialFailureKeepsSurvivorsAndShowsError() = runTest {
-        // Issue #570: when one of several images fails to upload, the ones
-        // that DID upload must still attach as tiles and the composer must
-        // re-enable (Idle) with a per-batch error — never discard everything
-        // or wedge on the failed file.
-        val vm = newVm()
-        vm.onDraftChange("Review these shots")
-        val first = "~/.pocketshell/attachments/host-1/20260606-120000-01-first.png"
-        val third = "~/.pocketshell/attachments/host-1/20260606-120000-03-third.png"
-
-        vm.attachFiles(count = 3) {
-            Result.failure(
-                PartialAttachmentUploadException(
-                    uploadedPaths = listOf(first, third),
-                    failedCount = 1,
-                    message = "Attached 2 of 3 files; 1 failed (upload timed out).",
-                ),
-            )
-        }
-        advanceUntilIdle()
-
-        // Survivors are attached; the failed file is absent.
-        assertEquals(
-            listOf(
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = first,
-                    displayName = "20260606-120000-01-first.png",
-                ),
-                PromptComposerViewModel.StagedAttachment(
-                    remotePath = third,
-                    displayName = "20260606-120000-03-third.png",
-                ),
-            ),
-            vm.uiState.value.attachments,
-        )
-        // Draft untouched, composer re-enabled, per-batch error visible.
-        assertEquals("Review these shots", vm.uiState.value.draft)
-        assertEquals(
-            PromptComposerViewModel.AttachmentUploadState.Idle,
-            vm.uiState.value.attachmentUpload,
-        )
-        val error = vm.uiState.value.error
-        assertNotNull(error)
-        assertTrue(error!!.contains("Attached 2 of 3"))
-        assertTrue(error.contains("1 failed"))
-    }
-
-    @Test
-    fun attachFilesInFlightDoesNotBlockDraftEdits() = runTest {
-        // Issue #570: staging/uploading attachments must not globally block
-        // text composition. The upload can stay busy, but the draft path
-        // remains editable and mirrored into saved state.
-        val vm = newVm()
-        val uploadStarted = CompletableDeferred<Unit>()
-        val uploadResult = CompletableDeferred<Result<List<String>>>()
-
-        vm.attachFiles(count = 3) {
-            uploadStarted.complete(Unit)
-            uploadResult.await()
-        }
-        runCurrent()
-        uploadStarted.await()
-
-        assertEquals(
-            PromptComposerViewModel.AttachmentUploadState.Uploading(3),
-            vm.uiState.value.attachmentUpload,
-        )
-
-        vm.onDraftChange("typing while images upload")
-
-        assertEquals("typing while images upload", vm.uiState.value.draft)
-        assertNull(vm.uiState.value.error)
-        assertEquals(
-            PromptComposerViewModel.AttachmentUploadState.Uploading(3),
-            vm.uiState.value.attachmentUpload,
-        )
-
-        uploadResult.complete(Result.success(emptyList()))
-        advanceUntilIdle()
-    }
-
-    @Test
-    fun attachFilesTimeoutClearsBusyStateAndKeepsDraft() = runTest {
-        val vm = newVm()
-        val uploadStarted = CompletableDeferred<Unit>()
-        vm.onDraftChange("keep this")
-
-        vm.attachFiles(count = 3) {
-            uploadStarted.complete(Unit)
-            awaitCancellation()
-        }
-        runCurrent()
-        uploadStarted.await()
-
-        advanceTimeBy(PromptComposerViewModel.ATTACHMENT_UPLOAD_TIMEOUT_MS)
-        runCurrent()
-
-        assertEquals("keep this", vm.uiState.value.draft)
-        assertEquals(
-            PromptComposerViewModel.AttachmentUploadState.Idle,
-            vm.uiState.value.attachmentUpload,
-        )
-        assertTrue(vm.uiState.value.attachments.isEmpty())
-        val error = vm.uiState.value.error
-        assertNotNull(error)
-        assertTrue(error!!.contains("Attachment upload failed"))
-        assertTrue(error.contains("timed out"))
-        assertTrue(error.contains("draft was kept"))
     }
 
     // -- Configured silence auto-stop ---------------------------------------
