+++ scripts/file-size-hygiene-baseline.txt
@@ -9,8 +9,8 @@
 180197 app/src/main/java/com/pocketshell/app/projects/FolderListViewModel.kt
 226939 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt
 941078 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
-272182 app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
-599334 app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
+258594 app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
+503704 app/src/test/java/com/pocketshell/app/tmux/TmuxSessionViewModelTest.kt
 1146350 docs/mockups/feedback/folder-tree-target-20260604.png
 182790 docs/screenshots/readme-settings.png
 194363 scripts/ci-journey-suite.sh
