+++ shared/ui-kit/src/test/java/com/pocketshell/uikit/render/DesignRenders.kt
@@ -711,39 +711,7 @@ class DesignRenders {
      */
     @Test
     fun hostListScreen() = render("host-list-screen") {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            ScreenHeader(
-                title = "Hosts",
-                subtitle = "5 hosts · 4 sessions",
-                trailing = {
-                    Badge(label = "4 active", role = BadgeRole.Active, mono = false)
-                },
-            )
-            HostCard(
-                name = "hetzner",
-                subtitle = "alex@65.108.42.11",
-                status = HostStatus.Attached,
-                onClick = {},
-            )
-            HostCard(
-                name = "gpu-box",
-                subtitle = "alex@10.0.0.42",
-                status = HostStatus.ActiveSessions(count = 3),
-                onClick = {},
-            )
-            HostCard(
-                name = "prod",
-                subtitle = "deploy@prod.acme.io",
-                status = HostStatus.NoActiveSessions,
-                onClick = {},
-            )
-            HostCard(
-                name = "edge",
-                subtitle = "ci@edge.acme.io",
-                status = HostStatus.ConnectionError,
-                onClick = {},
-            )
-        }
+        HostListScreenRender()
     }
 
     /**
@@ -772,44 +740,12 @@ class DesignRenders {
      */
     @Test
     fun hostCardResumeAffordance() = render("host-card-resume-affordance") {
-        Column(verticalArrangement = Arrangement.spacedBy(4.dp)) {
-            HostCard(
-                name = "hetzner",
-                subtitle = "alex@65.108.42.11",
-                status = HostStatus.Attached,
-                onClick = {},
-            )
-            ResumeLastSessionRowFacsimile(sessionName = "claude-main")
-        }
-        HostCard(
-            name = "gpu-box",
-            subtitle = "alex@10.0.0.42",
-            status = HostStatus.NoActiveSessions,
-            onClick = {},
-        )
+        HostCardResumeAffordanceRender()
     }
 
     @Test
     fun usageGlancePill() = render("usage-glance-pill") {
-        Column(verticalArrangement = Arrangement.spacedBy(16.dp)) {
-            ScreenHeader(
-                title = "Hosts",
-                subtitle = "5 hosts · 4 active",
-                trailing = {
-                    UsageGlancePillFacsimile(percent = "72%", dot = PillKind.Warn)
-                    AppBarPillFacsimile { Text("2", color = PocketShellColors.Text, style = PocketShellType.bodyDense, fontWeight = FontWeight.SemiBold) }
-                    AppBarGearFacsimile()
-                },
-            )
-            ScreenHeader(
-                title = "Hosts",
-                subtitle = "5 hosts · 4 active",
-                trailing = {
-                    UsageGlancePillFacsimile(percent = "63%", dot = PillKind.Ok, staleClock = "13:40")
-                    AppBarGearFacsimile()
-                },
-            )
-        }
+        UsageGlancePillRender()
     }
 
     /**
@@ -2690,113 +2626,6 @@ class DesignRenders {
                 text = text,
                 color = PocketShellColors.Text,
                 style = PocketShellType.bodyDense,
-            )
-        }
-    }
-
-    /**
-     * Issue #1241: ui-kit facsimile of the app-module app-bar usage glance pill
-     * ([com.pocketshell.app.usage.UsageGlancePill]) for the fast JVM render.
-     * Mirrors the app pill's chrome so the design read is faithful; the emulator
-     * screenshot is the acceptance.
-     */
-    @Composable
-    private fun UsageGlancePillFacsimile(percent: String, dot: PillKind, staleClock: String? = null) {
-        val alpha = if (staleClock != null) 0.6f else 1f
-        val dotColor = when (dot) {
-            PillKind.Ok -> PocketShellColors.Green
-            PillKind.Warn -> PocketShellColors.Amber
-            PillKind.Blocked -> PocketShellColors.Red
-            PillKind.Error -> PocketShellColors.TextMuted
-        }
-        Row(
-            modifier = Modifier
-                .height(32.dp)
-                .background(color = PocketShellColors.SurfaceElev, shape = PocketShellShapes.large)
-                .border(width = 1.dp, color = PocketShellColors.BorderSoft, shape = PocketShellShapes.large)
-                .padding(horizontal = 12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Box(modifier = Modifier.size(8.dp).clip(CircleShape).background(dotColor.copy(alpha = alpha)))
-            Spacer(modifier = Modifier.width(6.dp))
-            Text(
-                text = percent,
-                color = PocketShellColors.Text.copy(alpha = alpha),
-                style = PocketShellType.bodyDense,
-                fontWeight = FontWeight.SemiBold,
-            )
-            if (staleClock != null) {
-                Spacer(modifier = Modifier.width(6.dp))
-                Text(text = staleClock, color = PocketShellColors.TextMuted, style = PocketShellType.bodyDense)
-            }
-        }
-    }
-
-    @Composable
-    private fun AppBarPillFacsimile(content: @Composable RowScope.() -> Unit) {
-        Row(
-            modifier = Modifier
-                .height(32.dp)
-                .background(color = PocketShellColors.SurfaceElev, shape = PocketShellShapes.large)
-                .border(width = 1.dp, color = PocketShellColors.BorderSoft, shape = PocketShellShapes.large)
-                .padding(horizontal = 12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            content = content,
-        )
-    }
-
-    @Composable
-    private fun AppBarGearFacsimile() {
-        Box(
-            modifier = Modifier
-                .size(40.dp)
-                .clip(CircleShape)
-                .background(color = PocketShellColors.SurfaceElev)
-                .border(width = 1.dp, color = PocketShellColors.BorderSoft, shape = CircleShape),
-            contentAlignment = Alignment.Center,
-        ) {
-            Text("⚙", color = PocketShellColors.TextSecondary)
-        }
-    }
-
-    /**
-     * Issue #1239: JVM facsimile of the app-module `ResumeLastSessionRow` — the
-     * one-tap "Resume last session" affordance under the matching host card. Same
-     * chrome as the production row (accent play glyph, bright "Resume" label,
-     * muted-mono session name, AccentSoft fill + 40%-accent hairline on the
-     * `medium` card shape) so the fast render is a faithful design read; the
-     * emulator screenshot is the acceptance.
-     */
-    @Composable
-    private fun ResumeLastSessionRowFacsimile(sessionName: String) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(48.dp)
-                .background(color = PocketShellColors.AccentSoft, shape = PocketShellShapes.medium)
-                .border(
-                    width = 1.dp,
-                    color = PocketShellColors.Accent.copy(alpha = 0.4f),
-                    shape = PocketShellShapes.medium,
-                )
-                .padding(horizontal = 12.dp),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Text("▶", color = PocketShellColors.Accent, style = PocketShellType.bodyDense)
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = "Resume",
-                color = PocketShellColors.Accent,
-                style = PocketShellType.bodyDense,
-                fontWeight = FontWeight.SemiBold,
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = sessionName,
-                color = PocketShellColors.TextSecondary,
-                style = PocketShellType.bodyMono,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
             )
         }
     }
